requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.9.0
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==3.0.0
psutil==5.9.4
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.19.0
virtualenv-clone==0.5.7
